u
Fixed array of functions.  Changed and added variables for the array of
functions to avoid reusing variables.  Added finish3 function for return
finish and loop finish array functions.  Made the finishing statements
run when y is equal to as well as greater than or less than (depending
on the direction).  Note to self: May want to make this preventative, so
that y can never go beyond min or max.  Corrected array of functions
transfer to next function.